Consider invalid constituent trait when interpreting tokens
In the default readtable, the characters #\Backspace and #\Rubout have
constituent syntax type but also the invalid constituent trait. That
is, they are constituents when accumulating a token, but unless
escaped, they must cause an error to be signaled.

Other characters, such as #\Space, can exhibit the same behavior when
their syntax type is changed from whitespace to constituent.

fixes #35